+++ module-domain/src/main/java/com/zoopi/domain/phoneauthentication/entity/PhoneAuthenticationResult.java
@@ -1,6 +1,7 @@
-package com.zoopi.domain.phoneauthentication.dto.response;
+package com.zoopi.domain.phoneauthentication.entity;
 
 public enum PhoneAuthenticationResult {
+
 	AUTHENTICATED, NOT_AUTHENTICATED, EXPIRED, MISMATCHED
 
 }
